Migrating to Astro
Refactoring this site to be built on Astro. Additionally, playing around with a slightly newer
design, pulling in some AI content, and updating the respective social links. The concept is
still for a statically generated site, with any dynamic values being generated via build steps.

A set of todos would be to automate this build process via Actions.